+++ ci4/app/Interfaces/GatewayAnnotationList.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property GatewayAnnotation[] $values
+ */
+interface GatewayAnnotationList {
+
+}