+++ vue/src/core/services/Deploy/models/GatewayAnnotation.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 19-09-2026.
+ * Time: 16:08.
+ */
+import {GatewayAnnotationDefinition} from './definitions/GatewayAnnotationDefinition';
+
+export class GatewayAnnotation extends GatewayAnnotationDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}